fix(gradle-inspector): Optimize memory by caching dependency subtrees
In large projects with many dependencies, the GradleInspector plugin
often runs out of Java heap space, even when allocated 2 GB. This is
due to all dependencies being stored in a large tree structure, which
consumes a significant amount of memory.

This change introduces a caching mechanism for dependency subtrees of
components. Instead of traversing and storing the entire dependency
tree every time, the cache is checked first. If a subtree for a
component already exists in the cache, a reference to this cached
data is used, reducing memory consumption significantly.

Fixes #9220.

Signed-off-by: Wolfgang Klenk <wolfgang.klenk2@bosch.com>
(cherry picked from commit 9ccccf6b86b0f3aaf16add5067d8011b4511e33a)