feat(spdx): Add file level information to SPDX projects
In an SPDX report, the information of the project itself is converted to
an SPDX package entity as it is done for external dependencies as well.
However, this project entity does neither contain the attribute
`licenseInfoFromFiles` nor does it contain file-level information even
though the scanner found (and e.g. the web app report contains)
licenses. Only detected copyright statements are included in the field
`copyrightText`.

If the property `fileInformationEnabled` is set in the
`SPDXDocumentReporter`, file-level information for the scanned project
is included.

Co-authored-by: Kiko Fernandez-Reyes <kiko@erlang.org>
Co-authored-by: Daniel Kreck <daniel.kreck@collaboration-factory.de>

Signed-off-by: Kiko Fernandez-Reyes <kiko@erlang.org>

(cherry picked from commit 33b7a4f6784ad30674ef0348677595de153fb5ab)